changed weekday order
In my country (i.e., Brazil), weeks start on Sunday, always. However, in the rest of the world (except USA and some other places), weeks start on Monday, including the country of origin of Python. Because of that, I changed the order of weekdays in strings.en and strings.ptbr and changed `isoweekday() % 7` calls to only `weekday()`.